add header row support
this is a bit of a mess because each time we enumerate the file we have to re-read the header row. this is because position in bufferedstream is not the same as the position of the last character read so we can't store the offset of the header row after reading once and then seek back to it because we end up seeking back to wherever the buffer read up to rather than where the header row ended. we cache the header row value to avoid repeated seeks when accessing the header row property.